Late, Minor Commit
Made minor changes to the script itself. I'm toying around with a setup.py file to attempt to make
the project more readily accessible, but I really haven't put much effort into that because it is a
minor issue. Focus has been shifted to deciding on Natural Language Processing Algorithms, and once
we have selected an algorithm, programming will begin in earnest.